requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.4
celery==5.6.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==4.2.9
django-cors-headers==4.9.0
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
exceptiongroup==1.3.1
idna==3.11
inflection==0.5.1
kombu==5.6.1
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==7.1.0
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
wheel==0.45.1
gunicorn==21.2.0